added functionality to see if text has already been extracted or not and made some UI changes

+++ ParadymeIntern/RSR/views.py
@@ -101,8 +101,6 @@ def uploaddoc(request):
 
                 print (temp_doc.docfile.wordstr)
                 temp_doc.save(update_fields=['wordstr'])'''
-            temp_doc.docfile.wordstr=""
-            print (str(temp_doc.docfile))
             return HttpResponseRedirect(reverse('RSR:uploaddoc'))
     else:
         form = DocumentForm()
@@ -136,28 +134,32 @@ def listdelete(request, template_name='uploadlist.html'):
 
 @login_required
 def ocr (request):
-    temp_doc=Document()
-    docID=request.POST.get('docfile', None)
+    docID=request.POST.get('docfileID', None)
     documents=get_object_or_404(Document,pk=docID)
     var=str
-    if os.path.exists(str(settings.MEDIA_ROOT)+str(documents)):
-        status=True
-        file_path=str(settings.MEDIA_ROOT)+str(documents)
-        img=IMG(filename=file_path,resolution=200)
-        images=img.sequence
-        for i in range(len(images)):
-            IMG(images[i]).save(filename=str(settings.MEDIA_ROOT)+'/temp/'+str(i)+'.jpg')
-        for i in range(len(images)):
-            if i == 0:
-                var=get_string(str(settings.MEDIA_ROOT)+'/temp/'+str(i)+'.jpg')
-                os.remove(str(settings.MEDIA_ROOT)+'/temp/'+str(i)+'.jpg')
-            else:
-                var+="\n\n"
-                var+=get_string(str(settings.MEDIA_ROOT)+'/temp/'+str(i)+'.jpg')
-                os.remove(str(settings.MEDIA_ROOT)+'/temp/'+str(i)+'.jpg')
-        print(var)
+    if documents.wordstr in [None,""]:
+        if os.path.exists(str(settings.MEDIA_ROOT)+str(documents)):
+            status=True
+            file_path=str(settings.MEDIA_ROOT)+str(documents)
+            img=IMG(filename=file_path,resolution=200)
+            images=img.sequence
+            for i in range(len(images)):
+                IMG(images[i]).save(filename=str(settings.MEDIA_ROOT)+'/temp/'+str(i)+'.jpg')
+            for i in range(len(images)):
+                if i == 0:
+                    var=get_string(str(settings.MEDIA_ROOT)+'/temp/'+str(i)+'.jpg')
+                    os.remove(str(settings.MEDIA_ROOT)+'/temp/'+str(i)+'.jpg')
+                else:
+                    var+="\n\n"
+                    var+=get_string(str(settings.MEDIA_ROOT)+'/temp/'+str(i)+'.jpg')
+                    os.remove(str(settings.MEDIA_ROOT)+'/temp/'+str(i)+'.jpg')
+            Document.objects.filter(pk=docID).update(wordstr=var)
+
+        else:
+            status=False
     else:
-        status=False
+        status=True
+        var=str(documents.wordstr)
     context={'var':var, 'status':status}
     return render(request, 'ocr.html',context)
